app/requirements.txt
Django
django-filter
djangorestframework
mysqlclient
